+++ packages/core/shared/process-message-stream.ts
